pyproject.toml
[project]
name = "documents"
version = "3.0.9"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "aiohttp==3.10.2",
    "autodoc-pydantic~=1.8.0",
    "cachetools==5.5.0",
    "confluent-kafka==2.3.0",
    "fastapi==0.93.0",
    "flake8~=6.0.0",
    "grpcio==1.53.0",
    "minio~=7.1.9",
    "motor==3.1.1",
    "protobuf==3.20.3",
    "pydantic==1.10.7",
    "pyjwt[crypto]==2.6.0",
    "pymongo==4.3.3",
    "python-dateutil~=2.8.2",
    "python-multipart~=0.0.5",
    "sphinx~=6.1.3",
    "uvicorn==0.20.0",
]

[dependency-groups]
dev = [
]
lint = [
    "ruff==0.9.3",
]
tests = [
    "pytest~=7.2.2",
    "httpx==0.27.2",
    "requests==2.32.3",
]
security = [
    "pip-audit==2.7.3",
]
migrations = [
]
